Unify latency tests
