+++ tox.ini
@@ -61,7 +61,7 @@ commands =
     poetry run pytest \
         --benchmark-histogram={[main]latency_dumps_dir}/image \ 
         --benchmark-json={[main]latency_dumps_dir}/benchmark.json \
-        "orjson_benchmark/benchmark_dumps.py"
+        orjson_benchmark/benchmark_latency.py -k test_dumps
     poetry run graph-tabulate {[main]latency_dumps_dir}/benchmark.json
 
 [testenv:latency-empty]
@@ -71,7 +71,7 @@ commands =
     poetry run pytest \
         --benchmark-histogram={[main]latency_empty_dir}/image \
         --benchmark-json={[main]latency_empty_dir}/benchmark.json \ 
-        "orjson_benchmark/benchmark_empty.py"
+        orjson_benchmark/benchmark_latency.py -k test_empty
     poetry run graph-tabulate {[main]latency_empty_dir}/benchmark.json
 
 [testenv:latency-loads]
@@ -81,7 +81,7 @@ commands =
     poetry run pytest \
         --benchmark-histogram={[main]latency_loads_dir}/image \ 
         --benchmark-json={[main]latency_loads_dir}/benchmark.json \ 
-        "orjson_benchmark/benchmark_loads.py"
+        orjson_benchmark/benchmark_latency.py -k test_loads
     poetry run graph-tabulate {[main]latency_loads_dir}/benchmark.json
 
 [testenv:memory]
